+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -5,9 +5,9 @@ import ru.yandex.practicum.filmorate.model.Film;
 import java.util.List;
 
 public interface FilmService {
-	void liked(Integer filmId, Integer userId);
+	void liked(Long filmId, Long userId);
 
-	void disliked(Integer filmId, Integer userId);
+	void disliked(Long filmId, Long userId);
 
 	List<Film> getMostRatedFilms(int count);
